psalm fixes

+++ app/Models/P2pDailyStats.php
@@ -109,7 +109,7 @@ class P2pDailyStats extends Model
     /**
      * Accessor for total traffic
      */
-    public function total_traffic(): int
+    public function total_traffic(): float
     {
         return $this->ipv4_total_out + $this->ipv4_total_in + $this->ipv6_total_out + $this->ipv6_total_in;
     }
